refactored for clarity, moved substitutions into retry block
