+++ templates/shared_services/sonatype-nexus/template_schema.json
@@ -7,35 +7,111 @@
   "required": [],
   "properties": {},
   "pipeline": {
-    "upgrade": [
+    "install": [
       {
         "stepId": "main"
       },
       {
         "stepId": "a60421b3-7308-491f-b531-e007cdc0ff46",
-        "stepTitle": "Update the Firewall",
+        "stepTitle": "Add nexus rule collection to firewall",
         "resourceTemplateName": "tre-shared-service-firewall",
         "resourceType": "shared-service",
         "resourceAction": "upgrade",
         "properties": [
-        {
-          "name": "rule_collections",
-          "type": "array",
-          "substitutionAction": "overwrite",
-          "value": {
-            "name": "arc-web_app_subnet_nexus_api",
-            "action": "Allow",
-            "rules": [
+          {
+            "name": "rule_collections",
+            "type": "array",
+            "arraySubstitutionAction": "replace",
+            "arrayMatchField": "name",
+            "value": {
+              "name": "arc-web_app_subnet_nexus",
+              "action": "Allow",
+              "rules": [
                 {
-                    "name": "nexus-package-sources-api",
-                    "description": "Nexus Package Sources",
-                    "protocols": [{"port": "443", "type": "Https"}, {"port": "80", "type": "Http"}],
-                    "target_fqdns": "{{ resource.properties.nexus_allowed_fqdns_list }}",
-                    "source_addresses": "{{ resource.properties.address_prefixes }}"
+                  "name": "nexus-package-sources",
+                  "description": "Nexus Package Sources",
+                  "protocols": [
+                    {
+                      "port": "443",
+                      "type": "Https"
+                    },
+                    {
+                      "port": "80",
+                      "type": "Http"
+                    }
+                  ],
+                  "target_fqdns": "{{ resource.properties.nexus_allowed_fqdns_list }}",
+                  "source_addresses": "{{ resource.properties.address_prefixes }}"
                 }
-            ]
+              ]
+            }
           }
-        }]
+        ]
+      }
+    ],
+    "upgrade": [
+      {
+        "stepId": "main"
+      },
+      {
+        "stepId": "a60421b3-7308-491f-b531-e007cdc0ff47",
+        "stepTitle": "Update nexus rule collection in firewall",
+        "resourceTemplateName": "tre-shared-service-firewall",
+        "resourceType": "shared-service",
+        "resourceAction": "upgrade",
+        "properties": [
+          {
+            "name": "rule_collections",
+            "type": "array",
+            "arraySubstitutionAction": "replace",
+            "arrayMatchField": "name",
+            "value": {
+              "name": "arc-web_app_subnet_nexus",
+              "action": "Allow",
+              "rules": [
+                {
+                  "name": "nexus-package-sources",
+                  "description": "Nexus Package Sources",
+                  "protocols": [
+                    {
+                      "port": "443",
+                      "type": "Https"
+                    },
+                    {
+                      "port": "80",
+                      "type": "Http"
+                    }
+                  ],
+                  "target_fqdns": "{{ resource.properties.nexus_allowed_fqdns_list }}",
+                  "source_addresses": "{{ resource.properties.address_prefixes }}"
+                }
+              ]
+            }
+          }
+        ]
+      }
+    ],
+    "uninstall": [
+      {
+        "stepId": "a60421b3-7308-491f-b531-e007cdc0ff48",
+        "stepTitle": "Remove nexus rule collection from firewall",
+        "resourceTemplateName": "tre-shared-service-firewall",
+        "resourceType": "shared-service",
+        "resourceAction": "upgrade",
+        "properties": [
+          {
+            "name": "rule_collections",
+            "type": "array",
+            "arraySubstitutionAction": "remove",
+            "arrayMatchField": "name",
+            "value": {
+              "name": "arc-web_app_subnet_nexus"
+            }
+          }
+        ]
+      },
+      {
+        "stepId": "main"
       }
     ]
   }
